requirements.txt
# Core scientific computing packages
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0

# Molecular informatics
rdkit-pypi>=2022.3.1

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.60.0

# SHAP for model explainability
shap>=0.40.0

# Optional: for enhanced molecular visualization
py3Dmol>=1.8.0